Add French departements. Rename Centre.
Polygons were generated with polygons.openstreetmap.fr

The start positions are the coordinates of the préfectures (main
administrative cities of the départements). Coordinates are from
Wikidata.

Centre was renamed to "Centre-Val de Loire" by the law in 2015:
http://www.legifrance.gouv.fr/jopdf/common/jo_pdf.jsp?numJO=0&dateJO=20150117&numTexte=1&pageDebut=00777&pageFin=00783

Move French departements to their own xml.

Note: I have kept the same structure as example-config.xml so that
config-france.xml could be standalone.